既存jsをtsに置きかえ
